requirements.txt
github-copilot-sdk>=0.1.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
